Функция scaner() перенесена в extdecl
